# Vm apply uint256