--- a/vm/jump_table.go
+++ b/vm/jump_table.go
@@ -18,7 +18,6 @@ package vm
 
 import (
 	"errors"
-	"math/big"
 
 	"github.com/ethereum/go-ethereum/params"
 )
@@ -27,7 +26,7 @@ type (
 	executionFunc       func(pc *uint64, env *EVM, contract *Contract, memory *Memory, stack *Stack) ([]byte, error)
 	gasFunc             func(params.GasTable, *EVM, *Contract, *Stack, *Memory, uint64) (uint64, error) // last parameter is the requested memory size as a uint64
 	stackValidationFunc func(*Stack) error
-	memorySizeFunc      func(*Stack) *big.Int
+	memorySizeFunc      func(*Stack) (size uint64, overflow bool)
 )
 
 var errGasUintOverflow = errors.New("gas uint64 overflow")
@@ -453,7 +452,7 @@ func NewFrontierInstructionSet() JumpTable {
 			validateStack: makeStackFunc(0, 0),
 		},
 		PUSH1: {
-			execute:       makePush(1, 1),
+			execute:       opPush1,
 			gasCost:       gasPush,
 			validateStack: makeStackFunc(0, 1),
 		},