--- a/vm/logger.go
+++ b/vm/logger.go
@@ -147,8 +147,8 @@ func (l *StructLogger) CaptureState(env *EVM, pc uint64, op OpCode, gas, cost ui
 	// it in the local storage container.
 	if op == SSTORE && stack.len() >= 2 {
 		var (
-			value   = common.BigToHash(stack.data[stack.len()-2])
-			address = common.BigToHash(stack.data[stack.len()-1])
+			value   = stack.Back(1).Bytes32()
+			address = stack.Back(0).Bytes32()
 		)
 		l.changedValues[contract.Address()][address] = value
 	}
@@ -163,7 +163,7 @@ func (l *StructLogger) CaptureState(env *EVM, pc uint64, op OpCode, gas, cost ui
 	if !l.cfg.DisableStack {
 		stck = make([]*big.Int, len(stack.Data()))
 		for i, item := range stack.Data() {
-			stck[i] = new(big.Int).Set(item)
+			stck[i] = item.ToBig()
 		}
 	}
 	// Copy a snapshot of the current storage to a new container