--- a/vm/stack_table.go
+++ b/vm/stack_table.go
@@ -24,8 +24,8 @@ import (
 
 func makeStackFunc(pop, push int) stackValidationFunc {
 	return func(stack *Stack) error {
-		if err := stack.require(pop); err != nil {
-			return err
+		if stack.len() < pop {
+			return fmt.Errorf("stack underflow (%d <=> %d)", stack.len(), pop)
 		}
 
 		if stack.len()+push-pop > int(params.StackLimit) {