```
benchstat master.txt vm-apply-uint256.txt 
name            old time/op    new time/op    delta
OpAdd64-8         98.1ns ± 1%    15.0ns ± 0%   -84.67%  (p=0.008 n=5+5)
OpAdd128-8         111ns ± 0%      15ns ± 0%   -86.13%  (p=0.008 n=5+5)
OpAdd256-8        99.4ns ± 1%    15.1ns ± 1%   -84.80%  (p=0.008 n=5+5)
OpSub64-8         80.2ns ± 0%    12.9ns ± 0%   -83.86%  (p=0.000 n=4+5)
OpSub128-8        98.2ns ± 2%    12.8ns ± 0%   -86.95%  (p=0.016 n=5+4)
OpSub256-8         102ns ± 1%      15ns ± 0%   -85.18%  (p=0.016 n=5+4)
OpMul-8            137ns ± 1%      17ns ± 1%   -87.91%  (p=0.008 n=5+5)
OpDiv256-8         218ns ± 0%      48ns ± 2%   -77.72%  (p=0.016 n=4+5)
OpDiv128-8         122ns ± 1%      29ns ± 1%   -76.08%  (p=0.008 n=5+5)
OpDiv64-8         91.9ns ± 7%    14.9ns ± 1%   -83.83%  (p=0.008 n=5+5)
OpSdiv-8           327ns ± 2%      54ns ± 2%   -83.54%  (p=0.008 n=5+5)
OpMod-8            186ns ± 1%      18ns ± 0%   -90.41%  (p=0.008 n=5+5)
OpSmod-8           313ns ± 1%      23ns ± 1%   -92.68%  (p=0.008 n=5+5)
OpExp-8           11.7µs ± 0%     1.4µs ± 0%   -88.12%  (p=0.008 n=5+5)
OpSignExtend-8     104ns ± 1%      18ns ± 0%   -82.85%  (p=0.008 n=5+5)
OpLt-8            92.4ns ± 1%    15.4ns ± 1%   -83.36%  (p=0.008 n=5+5)
OpGt-8            92.8ns ± 1%    15.4ns ± 1%   -83.42%  (p=0.008 n=5+5)
OpSlt-8            102ns ± 0%      17ns ± 0%   -82.99%  (p=0.008 n=5+5)
OpSgt-8            100ns ± 1%      17ns ± 0%   -82.77%  (p=0.008 n=5+5)
OpEq-8            92.9ns ± 0%    15.4ns ± 1%   -83.44%  (p=0.008 n=5+5)
OpEq2-8           92.5ns ± 1%    15.1ns ± 1%   -83.71%  (p=0.008 n=5+5)
OpAnd-8           93.5ns ± 1%    15.2ns ± 0%   -83.76%  (p=0.008 n=5+5)
OpOr-8            93.5ns ± 1%    15.2ns ± 1%   -83.77%  (p=0.008 n=5+5)
OpXor-8           95.3ns ± 1%    15.3ns ± 2%   -83.95%  (p=0.008 n=5+5)
OpByte-8          90.6ns ± 1%    15.8ns ± 1%   -82.61%  (p=0.008 n=5+5)
OpAddmod-8         267ns ± 0%      69ns ± 1%   -74.04%  (p=0.008 n=5+5)
OpMulmod-8         352ns ± 0%     117ns ± 0%   -66.78%  (p=0.008 n=5+5)
OpSHL-8            114ns ± 1%      16ns ± 0%   -86.01%  (p=0.008 n=5+5)
OpSHR-8            110ns ± 0%      16ns ± 1%   -85.46%  (p=0.008 n=5+5)
OpSAR-8            186ns ± 0%      16ns ± 0%   -91.24%  (p=0.008 n=5+5)
OpIsZero-8        45.7ns ± 0%     8.7ns ± 1%   -81.05%  (p=0.008 n=5+5)

name            old alloc/op   new alloc/op   delta
OpAdd64-8           128B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpAdd128-8          184B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpAdd256-8          192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSub64-8          80.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
OpSub128-8          160B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSub256-8          192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpMul-8             288B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpDiv256-8          256B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpDiv128-8          168B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpDiv64-8          80.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
OpSdiv-8            416B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpMod-8             200B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSmod-8            424B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpExp-8           18.5kB ± 0%     0.0kB       -100.00%  (p=0.008 n=5+5)
OpSignExtend-8      200B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpLt-8              192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpGt-8              192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSlt-8             192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSgt-8             192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpEq-8              192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpEq2-8             192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpAnd-8             192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpOr-8              192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpXor-8             192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpByte-8            192B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpAddmod-8          336B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpMulmod-8          464B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSHL-8             136B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSHR-8             136B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpSAR-8             256B ± 0%        0B       -100.00%  (p=0.008 n=5+5)
OpIsZero-8         96.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name            old allocs/op  new allocs/op  delta
OpAdd64-8           5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpAdd128-8          5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpAdd256-8          4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSub64-8           4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSub128-8          4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSub256-8          4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpMul-8             5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpDiv256-8          5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpDiv128-8          5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpDiv64-8           4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSdiv-8            8.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpMod-8             5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSmod-8            9.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpExp-8              196 ± 0%         0       -100.00%  (p=0.008 n=5+5)
OpSignExtend-8      5.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpLt-8              4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpGt-8              4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSlt-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSgt-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpEq-8              4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpEq2-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpAnd-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpOr-8              4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpXor-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpByte-8            4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpAddmod-8          7.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpMulmod-8          8.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSHL-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSHR-8             4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpSAR-8             7.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
OpIsZero-8          2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```